Framework: refactor TransactionFlow
